Fix issue with OSX builds missing last line #84 (#85)
Removed dependancy on OSX default sed
Added steps to install gsed (GNU Sed) on OSX for development This gsed
on OSX is the same GNU sed used on GitHub Actions Possibly a different
version but generally the same features. This simplifies the Makefile
and avoids the duplicate effort to maintain.