Formatting improvements
Vendored in the markdownpp package so we can modify it to
work as we need, versus the previous approach of trying to
modify the output to match our requirements. This has greatly
reduced the complexity of sed operations and made it possible to
use css more effectively. Table of contents is now a HTML UL which
can be styled. Made some of the recommendations in Issue #89.