Update chargecategory.md
Parts of the initial definition are no longer applicable since changes to the allowed values list (credit is no longer part of the adjustment category but a separate category) and the introduction of additional columns. It has beeb raised for example, how this column would be used for separate charges that may require special handling from regular usage.