Fix broken builds reported in #438 (#439)
In order to unbreak the github actions pipeline, adding temporary fix
which will allow the pip install to apply to the system python on the
container. This is low risk as the python install is only active for the
life of the build. If we want to avoid this we need to do further
testing on using a venv for the spec creation. But this PR gets the
pipeline unblocked for the pending 1.0 release.

---------

Signed-off-by: Mike Fuller <mike@finops.org>