#441 ChargePeriod Start and End spec change (#442)
Based on #441 

Clarifying description and normative statements for expected values
between usage, time-bound purchase, and time-agnostic purchase scenarios

### v2

ChargePeriodStart

>ChargePeriodStart MUST be present in the billing data, MUST be of type
Date/Time, MUST be an *inclusive* value, and MUST NOT contain null
values. ChargePeriodStart MUST match the beginning date and time
boundary of the effective period of the charge.

ChargePeriodEnd

> ChargePeriodEnd MUST be present in the billing data, MUST be of type
Date/Time, MUST be an *exclusive* value, and MUST NOT contain null
values. ChargePeriodEnd MUST match the ending date and time boundary of
the effective period of the charge.


Note: there is a grey area when [ChargeCategory](#chargecategory) is
"Purchase" and the charge does not contain an expiration or ending date
and time boundary.

<details>
<summary>

### v1

</summary>

ChargePeriodStart

>ChargePeriodStart MUST be present in the billing data, MUST be of type
Date/Time, MUST be an *inclusive* value, and MUST NOT contain null
values. ChargePeriodStart MUST match the beginning date and time
boundary for which a charge is active.

ChargePeriodEnd

> ChargePeriodEnd MUST be present in the billing data, MUST be of type
Date/Time, MUST be an *exclusive* value, and MUST NOT contain null
values. When [ChargeCategory](#chargecategory) is "Purchase" and the
charge contains an expiration or ending date and time boundary,
ChargePeriodEnd MUST match the ending date and time boundary for which
the charge is active.

</details>