[ADMIN]update reference to patent policy 4 W3C mode as stated in the FOCUS Membership doc. (#480)
It was identified that the reference to the Patent Policy was incorrect
in the license document. It should be listed as Option 4, W3C mode,
stated in the FOCUS Membership Agreement rather than Patent Policy
Option 3: Open Web Foundation 1.0 Mode. See the screenshot below with
the correct reference to the W3C Mode patent.


![image](https://github.com/FinOps-Open-Cost-and-Usage-Spec/FOCUS_Spec/assets/3258579/25d47b58-7ae1-4a37-ad2a-443501ea1992)