[REPO] Enhance Issue Templates (#490)
### Enhance Template: `issue feedback `
This addition explicitly requests users to provide relevant data or data
extracts along with their feedback. It emphasizes the importance of data
in supporting discussions and decision-making while reminding submitters
to ensure that the data is anonymized and compliant with privacy
guidelines.

### Enhance Template: `discussion Topic`
This YAML replaces the previous `discussion-topic.md` file. It defines
the structure and content of the GitHub issue form for discussing broad
topics within the FOCUS community. It ensures that all necessary
information is gathered effectively, including an optional field for
attaching relevant data, which will support the discussion with tangible
examples.

### Enhance Template: `maintenance`
This YAML template is structured to gather essential information about
maintenance tasks within the GitHub repository or actions. The format
ensures clarity and specificity in the task descriptions and outcomes,
facilitating effective task management and completion.

### Enhance Template: `spec-change`
This template provides a YAML format for submitting specification
changes, ensuring that all relevant information is captured, including
the type of issue, whether the issue is normalized, a detailed
description, a clear definition of what constitutes completion of the
task, and an optional section for additional context or supporting
information. This approach ensures clarity and facilitates the efficient
management of specification changes.