Proposal for FOCUS v1.1 Scope (#504)
**Irena** and **Udam** have been working offline to formalize the scope
for FOCUS v1.1. This proposal is based on the current discussions and
Issues discussed in the three Task Forces (TF-1, TF-2, and TF-3).

In a nutshell, here is what Udam and I are proposing:

> _**SKU and SKU Price Details**_
> _Provide deeper visibility into SKU and SKU Price details (in addition
to the current SKU and SKU Price IDs that are available in FOCUS) which
enables many standard FinOps use cases from cost reporting to commitment
optimization._
>
>_**Deeper service classification**_
>_Move beyond the highest-level service categorization to a
sub-categorization for services while adding flexibility for providers
to share their native categorizations within FOCUS._
>
>_**Capacity Reservation**_
>_Ability to represent on-demand capacity reservations within FOCUS
datasets_
>
>_**Improved metadata support**_
>_Strengthen the correlation between FOCUS-compatible datasets and the
corresponding metadata to better support data analysis and ETL use
cases._

If everyone agrees, we could add that to the Release Planning as a
high-level scope for FOCUS 1.1.