#503a - Update SkuPriceId per SkuPriceDetails (#530)
This change is dependent on #503 being merged as it links to
SkuPriceDetails. This change to SkuPriceId solves for a case where a
provider does not have a SkuPriceId but wants to include
SkuPriceDetails. This case was introduced when there was a change in
scope of #503 to not include SkuDetails.

Changes:
- Reformatted normative to bullets
- Added option (MAY) to repeat SkuId as SkuPriceId when there is no
SkuPriceId, as long as other conditions are not violated.

---------

Co-authored-by: Udam Dewaraja <udam@finops.org>
Co-authored-by: Joaquin <jprado@linuxfoundation.org>