billingcurrency normative requirement change
Change billing Currency normative requirement of currency matching invoice currency from MUST to SHOULD as AWS billing currency is USD only in FOCUS and CUR exports but invoices may be in any other currency of the customers choosing. The Description also insists that the column represents the currency the charge was billed in and therefore we shouldn't be enforcing the requirement to match currency in the dataset and the invoice.