AI #872: Revised 1. Introduction (#878)
This PR is related to Action Item #872

Detailed Description
Issue

The first paragraph in section 1. Introduction should be enhanced to
make it clear that FOCUS is a Standards Development Organization (SDO).

"FOCUS aims to establish a community-driven specification for
consumption-based billing data. ..."
Suggestion under Section 1. Introduction

Replace that statement for something like:

" FOCUS is a standards development organization (SDO) formed to
establish an open, consensus-driven standard for consumption-based
billing data. In the absence of a broadly adopted standard,
infrastructure and service providers have resorted to proprietary
billing schemas and inconsistent terminology, making cost data difficult
to normalize and act upon across environments.

The FOCUS Specification is developed by a global community of
practitioners and vendors working collaboratively to define a
consistent, vendor-neutral approach to billing data. It is designed to
improve interoperability, reduce complexity, and enable transparency in
cloud and SaaS cost management."

---------

Co-authored-by: Mike Fuller <mike@finops.org>
Co-authored-by: Matt Cowsert <matthew@finops.org>