Update invoice_handling.md
Took a stab at the scope of charges and tying the costs to the billing account as defined in the glossary. I also added some additional clarification for informational/supplemental details that often end up on the invoice but aren't expected to be present as separate line items in a FOCUS Cost and Usage dataset.