AI #1530: Revise Status values (#1585)
We have added a new entry to the Status attribute of the FOCUS WG
project view (`In Evaluation`) to represent the Feature Request Review
process we have introduced to FOCUS 1.4.

This PR updates the documentation for the Status attribute in the
development processes guidelines.

This PR is NOT to be approved and merged as part of 1.3, as it was not
part of that release process. We can approve and merge in any Members
meeting after Nov 6.

---------

Co-authored-by: Joaquin Prado <joaquin@elastichub.io>
Co-authored-by: Joaquin <jprado@linuxfoundation.org>